% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fs_estep_cpp}
\alias{fs_estep_cpp}
\title{Full-sib E-step: observed log-likelihood and expected maternal/paternal switch
and total counts per interval.}
\usage{
fs_estep_cpp(G, Am, Ap, rm, rp, epsilon, return_gamma = FALSE)
}
\description{
Full-sib E-step: observed log-likelihood and expected maternal/paternal switch
and total counts per interval.
}
\keyword{internal}
