% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fs_loglik_cpp}
\alias{fs_loglik_cpp}
\title{Full-sib four-state forward log-likelihood (summed over offspring)}
\usage{
fs_loglik_cpp(G, Am, Ap, rm, rp, epsilon)
}
\description{
Full-sib four-state forward log-likelihood (summed over offspring)
}
\keyword{internal}
