% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_utils.R
\name{hs_pal}
\alias{hs_pal}
\title{Palette for HSMap dendrograms}
\usage{
hs_pal(n)
}
\arguments{
\item{n}{Positive integer number of groups}
}
\value{
Character vector of length \code{n} with hex colors
}
\description{
Palette for HSMap dendrograms
}
