% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_ripple.R
\name{plot.HSMap.try}
\alias{plot.HSMap.try}
\title{Plot a marker placement profile}
\usage{
\method{plot}{HSMap.try}(x, ...)
}
\arguments{
\item{x}{An \code{HSMap.try} object.}

\item{...}{Passed to \code{\link[graphics]{plot}}.}
}
\value{
\code{x}, invisibly.
}
\description{
Plot a marker placement profile
}
