% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HSMap-package.R, R/zzz.R
\docType{package}
\name{HSMap-package}
\alias{HSMap-package}
\title{momap package}
\description{
A fast C++ (Rcpp) implementation that uses a two-state Hidden Markov Model to estimate recombination rates between genetic markers from offspring genotypes in outcrossing diploid species. It handles known mothers, optional known sires, multiple populations (families), and scattered missing markers.

A fast C++ (Rcpp) implementation that uses a two-state Hidden Markov Model to estimate recombination rates between genetic markers from offspring genotypes in outcrossing diploid species. It handles known mothers, optional known sires, multiple populations (families), and scattered missing markers.
}
\author{
\strong{Maintainer}: Marcelo Mollinari \email{mmollina@gmail.com}

}
\keyword{internal}
